requirements.txt.txt
PyNaCl
pytube
discord.py==1.7.3
ffmpeg
